Following bugfix
Fixed an error where it would follow a random user if they retweeted a "retweet to win tweet" with the word follow in it, and wouldn't follow the actual contest-holder. This fixes that, and also an error where the program crashes if it tries to retweet something it's already retweeted.