Update twitterbot.py
The bot was entering a bunch of crap like "Justin beiber is the hottest! Retweet to vote for him!", so i added the word "vote" to the bannedlist. If you can think of anything else that needs to be put on the banned-word list create an issue and i'll put it in.